+++ models/index.ts
@@ -1,0 +1,11 @@
+export type userType = {
+  id: string;
+  name: string;
+  age: number;
+  isAdmin: boolean;
+  experience: number;
+};
+
+export type paramsType = {
+  id: string;
+};
